backend/requirements.txt
<<<<<<< HEAD
# Reuse the main production requirements list
-r requirements_production.txt 
=======
flask==2.3.3
flask-socketio==5.3.6
gunicorn==21.2.0
flask-cors==4.0.0
python-socketio==5.8.0
requests==2.31.0
redis==4.6.0
python-dotenv==1.0.0
supabase==2.0.2
cryptography==41.0.4
openai==1.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
stripe==7.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0 
>>>>>>> 2a86997 (Add frontend fallback authentication processing)
